+++ src/react-components/room/InvitePopover.js
@@ -8,6 +8,7 @@ import { ReactComponent as InviteIcon } from "../icons/Invite.svg";
 import { Column } from "../layout/Column";
 import { InviteLinkInputField } from "./InviteLinkInputField";
 import { FormattedMessage, defineMessage, useIntl } from "react-intl";
+import { ToolTip } from "@mozilla/lilypad-ui";
 
 function InvitePopoverContent({ url, embed, inviteRequired, fetchingInvite, inviteUrl, revokeInvite }) {
   return (
@@ -45,6 +46,11 @@ InvitePopoverContent.propTypes = {
   revokeInvite: PropTypes.func
 };
 
+const inviteTooltipDescription = defineMessage({
+  id: "invite-tooltip.description",
+  defaultMessage: "Copy room link to invite others to the room"
+});
+
 const invitePopoverTitle = defineMessage({
   id: "invite-popover.title",
   defaultMessage: "Invite"
@@ -63,6 +69,7 @@ export function InvitePopoverButton({
 }) {
   const intl = useIntl();
   const title = intl.formatMessage(invitePopoverTitle);
+  const description = intl.formatMessage(inviteTooltipDescription);
 
   return (
     <Popover
@@ -83,14 +90,16 @@ export function InvitePopoverButton({
       popoverApiRef={popoverApiRef}
     >
       {({ togglePopover, popoverVisible, triggerRef }) => (
-        <ToolbarButton
-          ref={triggerRef}
-          icon={<InviteIcon />}
-          selected={popoverVisible}
-          onClick={togglePopover}
-          label={title}
-          {...rest}
-        />
+        <ToolTip description={description}>
+          <ToolbarButton
+            ref={triggerRef}
+            icon={<InviteIcon />}
+            selected={popoverVisible}
+            onClick={togglePopover}
+            label={title}
+            {...rest}
+          />
+        </ToolTip>
       )}
     </Popover>
   );
