+++ src/react-components/room/PlacePopover.js
@@ -5,6 +5,12 @@ import { Popover } from "../popover/Popover";
 import { ToolbarButton } from "../input/ToolbarButton";
 import { ReactComponent as ObjectIcon } from "../icons/Object.svg";
 import { defineMessage, useIntl } from "react-intl";
+import { ToolTip } from "@mozilla/lilypad-ui";
+
+const placeTooltipDescription = defineMessage({
+  id: "place-tooltip.description",
+  defaultMessage: "Select from a variety of objects and tools to edit your room"
+});
 
 const placePopoverTitle = defineMessage({
   id: "place-popover.title",
@@ -21,6 +27,7 @@ export function PlacePopoverButton({ items }) {
   }
 
   const title = intl.formatMessage(placePopoverTitle);
+  const description = intl.formatMessage(placeTooltipDescription);
 
   return (
     <Popover
@@ -30,14 +37,16 @@ export function PlacePopoverButton({ items }) {
       offsetDistance={28}
     >
       {({ togglePopover, popoverVisible, triggerRef }) => (
-        <ToolbarButton
-          ref={triggerRef}
-          icon={<ObjectIcon />}
-          selected={popoverVisible}
-          onClick={togglePopover}
-          label={title}
-          preset="accent3"
-        />
+        <ToolTip description={description}>
+          <ToolbarButton
+            ref={triggerRef}
+            icon={<ObjectIcon />}
+            selected={popoverVisible}
+            onClick={togglePopover}
+            label={title}
+            preset="accent3"
+          />
+        </ToolTip>
       )}
     </Popover>
   );
