+++ src/react-components/room/SharePopover.js
@@ -5,6 +5,12 @@ import { Popover } from "../popover/Popover";
 import { ToolbarButton } from "../input/ToolbarButton";
 import { ReactComponent as ShareIcon } from "../icons/Share.svg";
 import { defineMessage, useIntl } from "react-intl";
+import { ToolTip } from "@mozilla/lilypad-ui";
+
+const shareTooltipDescription = defineMessage({
+  id: "share-tooltip.description",
+  defaultMessage: "Display your screen or webcam as an object in the room"
+});
 
 const sharePopoverTitle = defineMessage({
   id: "share-popover.title",
@@ -14,6 +20,7 @@ const sharePopoverTitle = defineMessage({
 export function SharePopoverButton({ items }) {
   const intl = useIntl();
   const title = intl.formatMessage(sharePopoverTitle);
+  const description = intl.formatMessage(shareTooltipDescription);
 
   const filteredItems = items.filter(item => !!item);
 
@@ -52,14 +59,16 @@ export function SharePopoverButton({ items }) {
       disableFullscreen
     >
       {({ togglePopover, popoverVisible, triggerRef }) => (
-        <ToolbarButton
-          ref={triggerRef}
-          icon={<ShareIcon />}
-          selected={popoverVisible}
-          onClick={togglePopover}
-          label={title}
-          preset="accent5"
-        />
+        <ToolTip description={description}>
+          <ToolbarButton
+            ref={triggerRef}
+            icon={<ShareIcon />}
+            selected={popoverVisible}
+            onClick={togglePopover}
+            label={title}
+            preset="accent5"
+          />
+        </ToolTip>
       )}
     </Popover>
   );
