+++ src/react-components/room/SharePopoverContainer.js
@@ -4,15 +4,9 @@ import { ReactComponent as VideoIcon } from "../icons/Video.svg";
 import { ReactComponent as DesktopIcon } from "../icons/Desktop.svg";
 import { ReactComponent as AvatarIcon } from "../icons/Avatar.svg";
 import { SharePopoverButton } from "./SharePopover";
-import { FormattedMessage, defineMessage, useIntl } from "react-intl";
+import { FormattedMessage } from "react-intl";
 import useAvatar from "./hooks/useAvatar";
 import { MediaDevicesEvents, MediaDevices } from "../../utils/media-devices-utils";
-import { ToolTip } from "@mozilla/lilypad-ui";
-
-const shareTooltipDescription = defineMessage({
-  id: "share-tooltip.description",
-  defaultMessage: "Display your screen or webcam as an object in the room"
-});
 
 function useShare(scene, hubChannel) {
   const mediaDevicesManager = APP.mediaDevicesManager;
@@ -116,8 +110,6 @@ function useShare(scene, hubChannel) {
 }
 
 export function SharePopoverContainer({ scene, hubChannel }) {
-  const intl = useIntl();
-  const description = intl.formatMessage(shareTooltipDescription);
   const {
     sharingSource,
     canShareCamera,
@@ -155,11 +147,7 @@ export function SharePopoverContainer({ scene, hubChannel }) {
     }
   ];
 
-  return (
-    <ToolTip description={description}>
-      <SharePopoverButton items={items} />
-    </ToolTip>
-  );
+  return <SharePopoverButton items={items} />;
 }
 
 SharePopoverContainer.propTypes = {
